% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_kernel.R
\name{add_kernel}
\alias{add_kernel}
\title{Add a new kernel to the IPM}
\usage{
add_kernel(proto_ipm, name, formula, family, ..., data_list, state_list,
  dom_start, dom_end, int_rule = "midpoint", evict = TRUE,
  evict_type = "truncated_distributions")
}
\arguments{
\item{proto_ipm}{The name of object you wish to append the new kernel to.}

\item{name}{The name of the new kernel. Survival and growth kernels should be
named \code{"P"} for now. Fecundity should be named \code{"F"}, and clonal
kernels should be named \code{"C"} (again, just for now. \strong{ need to work on
generalizing that).}}

\item{formula}{An bare expression specifying the form of the kernel. See below
for examples}

\item{family}{The type of kernel. Options are \code{"CC"} for continuous to continuous
transitions, \code{"DC"} for discrete to continuous (e.g. emergence from a seedbank),
\code{"CD"} for continuous to discrete (e.g. entering a seedbank), and \code{"DD"} for
discrete to discrete (e.g. stasis in a seedbank).}

\item{...}{A set of named expressions that correspond to vital rates in \code{formula}.}

\item{data_list}{A list of named values that correspond to constants in the formula.
You do not need to specify vectors corresponding to the domains here.}

\item{state_list}{A list, usually the output of \code{define_state_vars}. If specified
manually, then each entry of the list should be a vector of length 4. The first is
the name of the state variable (e.g. "dbh", "height", "weight", etc), the second
is the lower boundary for that state variable (e.g. 2, 0.5, etc), the 3rd the upper
bound, and the 4th the number of meshpoints for that particular state variable.}

\item{dom_start}{The name of the state variable for the kernel at time \emph{t}.}

\item{dom_end}{The name of the state variable for the kernel at time \emph{t+1}.
This is usually the same as \code{dom_start}, but general IPMs
with discrete classes or IPMs that move from one state to another may have another
value here. For cases with a discrete stage, kernels moving individuals from
discrete to continuous should have a state variable entered here and an \code{NA}
for \code{dom_start}. For kernels moving from continuous to discrete, vice versa.
For discrete to discrete, both are \code{NA}.}

\item{int_rule}{The integration rule to be used for the kernel. The default is
"midpoint". "trapezoid" and "g-l" (Gauss-Legendre) will be implemented as well.
If "g-l", additional arguments need to be supplied (\strong{Work on this later!!}).}

\item{evict}{A logical indicating whether an eviction correction should be applied
to the kernel. Default is \code{TRUE}.}

\item{evict_type}{If \code{evict == TRUE}, then the type of eviction. Currently,
\code{"truncated_distributions"} is the default (and the only option for now!).}
}
\description{
Adds a new kernel to the \code{proto_ipm} structure. Different
classes of IPMs may have many or only a few kernels. Each one requires its
own call to this function, though there are some exceptions, namely for kernels
derived from hierarchical models (e.g. vital rate models fit across plots and years).
See Details for more info
}
\details{
\strong{BLAH BLAH BLAH}
}
