% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-age_size.R
\name{sum.age_size_expr}
\alias{sum.age_size_expr}
\alias{prod.age_size_expr}
\title{Age X Size models}
\usage{
\method{sum}{age_size_expr}(expr, ..., na.rm = NA)

\method{prod}{age_size_expr}(expr, ..., na.rm = NA)
}
\arguments{
\item{expr}{An expression to compute a sum for.}

\item{...}{Used internally. Modifying this will likely cause models to break.}

\item{na.rm}{Not used and ignored.}
}
\value{
A modified expression.
}
\description{
Helper functions for implementing age x size models. These are
only intended for use within \code{define_kernel()}.
}
\details{
Note that these functions do not compute a single number, but rather
expand an expression to compute the trait distribution. This differs from the
usual behavior of the numeric methods for \code{sum, prod} as those
return a single number or \code{NA}.
}
