% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_kernel.R
\name{define_kernel}
\alias{define_kernel}
\title{Add a new kernel to the IPM}
\usage{
define_kernel(proto_ipm, name, formula, family, ..., data_list = list(),
  state_list, dom_start, dom_end, int_rule = "midpoint",
  has_hier_effs = FALSE, levels_hier_effs = list(), evict = TRUE,
  evict_fun = NULL)
}
\arguments{
\item{proto_ipm}{The name of object you wish to append the new kernel to.}

\item{name}{The name of the new kernel. Survival and growth kernels should be
named \code{"P"} for now. Fecundity should be named \code{"F"}, and clonal
kernels should be named \code{"C"} (again, just for now. \strong{ need to work on
generalizing that).}}

\item{formula}{An bare expression specifying the form of the kernel. See below
for examples}

\item{family}{The type of kernel. Options are \code{"CC"} for continuous to continuous
transitions, \code{"DC"} for discrete to continuous (e.g. emergence from a seedbank),
\code{"CD"} for continuous to discrete (e.g. entering a seedbank), and \code{"DD"} for
discrete to discrete (e.g. stasis in a seedbank).}

\item{...}{A set of named expressions that correspond to vital rates in \code{formula}.}

\item{data_list}{A list of named values that correspond to constants in the formula.
You do not need to specify vectors corresponding to the domains here.}

\item{state_list}{A list containing the names of each state variable used in
the kernel.}

\item{dom_start}{The name of the state variable for the kernel at time \emph{t}.}

\item{dom_end}{The name of the state variable for the kernel at time \emph{t+1}.
This is usually the same as \code{dom_start}, but general IPMs
with discrete classes or IPMs that move from one state to another (e.g. tree
seedling going from a height domain to a DBH domain at T+1) may have another
value here. For cases with a discrete stage, kernels moving individuals from
discrete to continuous should have a state variable entered here and an \code{NA}
for \code{dom_start}. For kernels moving from continuous to discrete, vice versa.
For discrete to discrete, both are \code{NA}.}

\item{int_rule}{The integration rule to be used for the kernel. The default is
"midpoint". "trapezoid" and "g-l" (Gauss-Legendre) will be implemented as well.
If "g-l", additional arguments need to be supplied (\strong{Work on this later!!}).}

\item{has_hier_effs}{A logical indicating whether or not the kernel and/or its
underlying vital rates are structured with hierarchical effects. If so and you
specify either the functional forms or the exact parameter values you want to use,
the kernels will be automatically split according to the notation in the vital
rates and kernel formulae and multiple kernels will be built. See the vignette
on the syntax for this feature for more details (\code{vignettes(
'hierarchical-notation', package = 'ipmr')}).}

\item{evict}{A logical indicating whether an eviction correction should be applied
to the kernel. Default is \code{TRUE}.}

\item{evict_fun}{If \code{evict == TRUE}, then a function that corrects for it.
Currently, \code{truncated_distributions} and \code{rescale_kernel} are the
only built-in functions, but user specified ones should work as well.}
}
\description{
Adds a new kernel to the \code{proto_ipm} structure. Different
classes of IPMs may have many or only a few kernels. Each one requires its
own call to this function, though there are some exceptions, namely for kernels
derived from hierarchical models (e.g. vital rate models fit across plots and years).
See Details for more info
}
\details{
\strong{BLAH BLAH BLAH}
}
