% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{elasticity}
\alias{elasticity}
\title{Compute elasticity}
\usage{
elasticity(
  ipm,
  level = c("kernel", "vital_rate", "parameter"),
  subset = NA_character_,
  ...
)
}
\arguments{
\item{ipm}{Output from \code{make_ipm()}.}

\item{level}{The level to compute elasticity at. \code{"kernel"} computes
the model wide elasticity surface and returns that. \code{"vital_rate"}
computes lambda's elasticity to specific vital rates, \code{"parameter"} lambda's
elasticity to each parameter. Use \code{subset} to specify a subset of
vital rates or parameters to compute values for.}

\item{subset}{A character vector corresponding to the \code{"vital_rates"} or
\code{"parameters"} you want to compute  sensitivities for. Can save time if
only a few parameters or vital rates are of interest.}

\item{...}{other arguments passed to methods}
}
\value{
A list. Contains either kernel response surfaces, vital rate level
elasticity values, or parameter level elasticity values. The class is always
\code{c("ipmr_elasticity", "list")} (for internal usage and plot methods).
}
\description{
Compute elasticity
}
