% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-export.R
\name{right_mult}
\alias{right_mult}
\alias{s_g_mult}
\title{Function multiplication helpers}
\usage{
right_mult(...)

s_g_mult(s, g)
}
\arguments{
\item{...}{Symbols representing functions - usually things that appear on the
left hand side of \code{formula} and/or \code{...} in \code{define_kernel} and
\code{define_k}.}

\item{s}{A vector of survival probabilities}

\item{g}{A discretized growth kernel}
}
\description{
Helpers for multiplying functions - follows the terminology from
Ellner, Rees & Childs (2016), Table 3.1. \code{s_g_mult} is a special function
that correctly multiplies survival and growth
}
