% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{gen_di_det_ex}
\alias{gen_di_det_ex}
\title{A general deterministic IPM example}
\format{
A general deterministic IPM with the following slots:
\describe{

  \item{sub_kernels}{The computed sub-kernels for the model, named \code{P},
  \code{go_discrete}, \code{stay_discrete}, and \code{leave_discrete}.}
  \item{env_list}{Empty}
  \item{env_seq}{Contains \code{NA}. Not particularly useful for deterministic IPMs,
  but critical for reproducing stochastic ones.}
  \item{pop_state}{A list of length 2, with names \code{n_b} and
  \code{n_ht}.}
  \item{proto_ipm}{The \code{proto_ipm} used to implement the model.}

}
}
\usage{
gen_di_det_ex
}
\description{
A general deterministic IPM example
}
\keyword{datasets}
