% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-export.R
\name{int_mesh}
\alias{int_mesh}
\title{Accessor functions for IPM objects}
\usage{
int_mesh(ipm)
}
\arguments{
\item{ipm}{An object created by \code{make_ipm}.}
}
\value{
A list. Names correspond to the state variables, and values are the
domains. There will also be entries named \code{"d_<stateVariable>"} which
are the integration weights for a given domain.
}
\description{
Functions that access specific slots of a \code{*_ipm} for
user examination. These are not the same as the accessors for a \code{proto_ipm}.
}
