% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-export.R
\name{ipm_to_df}
\alias{ipm_to_df}
\title{Convert ipmr matrix to long data frame}
\usage{
ipm_to_df(ipm, ...)
}
\arguments{
\item{ipm}{Output from \code{make_ipm}.}

\item{...}{Other arguments passed to methods.}
}
\value{
A data frame with 3 columns named \code{"t"}, \code{"t_1"}, and
\code{"value"}.
}
\description{
Converts IPM kernels into long data frames. These are useful for
creating plots using \code{ggplot2}.
}
\examples{

data(gen_di_det_ex)

big_mat_df <- ipm_to_df(gen_di_det_ex,
                        mega_mat = c(stay_discrete, go_discrete,
                                     leave_discrete, P))

}
