% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{lambda}
\alias{lambda}
\alias{lambda.simple_di_det_ipm}
\alias{lambda.simple_di_stoch_kern_ipm}
\alias{lambda.simple_di_stoch_param_ipm}
\alias{lambda.general_di_det_ipm}
\alias{lambda.general_di_stoch_kern_ipm}
\alias{lambda.general_di_stoch_param_ipm}
\title{Compute the per-capita growth rate for an IPM object}
\usage{
lambda(ipm, ...)

\method{lambda}{simple_di_det_ipm}(ipm, comp_method = c("eigen", "pop_size"), all_lambdas = TRUE, ...)

\method{lambda}{simple_di_stoch_kern_ipm}(ipm, comp_method = c("eigen", "pop_size"), all_lambdas = TRUE, ...)

\method{lambda}{simple_di_stoch_param_ipm}(ipm, comp_method = c("eigen", "pop_size"), all_lambdas = TRUE, ...)

\method{lambda}{general_di_det_ipm}(ipm, all_lambdas = TRUE, ...)

\method{lambda}{general_di_stoch_kern_ipm}(ipm, ..., all_lambdas = TRUE)

\method{lambda}{general_di_stoch_param_ipm}(ipm, ..., all_lambdas = TRUE)
}
\arguments{
\item{ipm}{An object returned by \code{make_ipm()}.}

\item{...}{other arguments passed to methods.}

\item{comp_method}{Either \code{"eigen"} or \code{"pop_size"}. \code{"eigen"}
is only possible for \code{"simple_*"} methods.}

\item{all_lambdas}{A logical: return lambdas for each individual kernel, or
a single number? Only applies to \code{comp_method = "pop_size"}. When
\code{TRUE}, returns the ratio of population sizes for every iteration of the
model. When \code{FALSE}, only returns the ratio of population sizes for the
final iteration.}
}
\value{
A single numeric vector or single value.
}
\description{
Compute the per-capita growth rate for an IPM object
}
\details{
There are two possible methods for computing \code{lambda} and these
are controlled by the \code{comp_method} argument. Possible values are
\code{"eigen"} and \code{"pop_size"}. The first computes the dominant
eigenvalue of all entries in \code{ipm$iterators} slot of the \code{*_ipm}
object. Since iteration kernels aren't generated for \code{general_*} methods,
\code{"pop_size"} is the only possible option for those objects.
}
