% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-export.R
\name{make_iter_kernel}
\alias{make_iter_kernel}
\title{Create iteration kernels from an IPM object}
\usage{
make_iter_kernel(ipm, mega_mat = NULL, name_ps = NULL, f_forms = NULL)
}
\arguments{
\item{ipm}{Output from \code{make_ipm}.}

\item{mega_mat}{A vector with symbols, I's, and/or 0s representing the matrix blocks.
They should be specified in ROW MAJOR order! Can also be a character
string specifying the call. Hierarchical syntax is supported. When used,
\code{format_mega_matrix} will produce as many mega-matrices as there are
combinations of \code{levels_hier_effs} in the \code{proto_ipm}.}

\item{name_ps}{The prefix(es) for the kernel name that correspond to survival
and growth/maturation of existing individuals. For the model
\code{K = P_age + F_age}, this would be \code{"P"}.}

\item{f_forms}{The names of the kernels that correspond to production of new
individuals, and possibly, how they are combined. For example, a model that
includes sexual (with an "F" kernel) and asexual reproduction (with a "C" kernel),
this would be \code{"F + C"}. If data come from multiple sites or years,
then this information is supplied using the suffix syntax (i.e.
\code{f_forms = "F_yr + C_yr"}).}
}
\description{
Creates iteration kernels for IPMs. \code{ipmr} does not create
these to iterate models, but they may be useful for further analyses.
}
\details{
\code{ipmr} does not generate complete iteration kernels, and uses
sub-kernels to iterate models. However, some further analyses are just easier
to code with a complete iteration kernel. This handles constructing those for
simple and general models of all forms. For general models, this is really
just a thin wrapper around \code{format_mega_matrix}.
}
