% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as.matrix.ipmr_matrix}
\alias{as.matrix.ipmr_matrix}
\alias{as.matrix.ipmr_ipm}
\title{Convert to bare matrices}
\usage{
\method{as.matrix}{ipmr_matrix}(x, ...)

\method{as.matrix}{ipmr_ipm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ipmr_matrix}, or the output from
\code{make_ipm}.}

\item{...}{ignored.}
}
\value{
A matrix.
}
\description{
Converts objects to \code{c("matrix", "array")}.
}
