% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-export.R
\name{\%^\%}
\alias{\%^\%}
\alias{mat_power}
\title{Raise a matrix to a power}
\usage{
x \%^\% y

mat_power(x, y)
}
\arguments{
\item{x}{A numeric or integer matrix.}

\item{y}{An integer.}
}
\value{
A matrix.
}
\description{
Raises a matrix \code{x} to the \code{y}-th power. \code{x ^ y} computes
element wise powers, whereas this computes \emph{y - 1} matrix multiplications.
\code{mat_power(x, y)} is identical to \code{x \%^\% y}.
}
