% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stoch.R
\name{mean_kernel}
\alias{mean_kernel}
\title{Mean kernels for stochastic models}
\usage{
mean_kernel(ipm)
}
\arguments{
\item{ipm}{A stochastic model created by \code{make_ipm()}.}
}
\value{
A list of mean sub-kernels for the model.
}
\description{
This function computes mean sub-kernels for stochastic
parameter re-sampled and stochastic kernel re-sampled models.
}
\details{
For \code{*_stoch_kern} models, this computes the element-wise
mean for each sub-kernel across all the different levels of
\code{levels_hier_effs}. For partially hierarchical models, sub-kernels that
do not have time- or space-varying components are included in the output and
are identical to their input.

For \code{*_stoch_param} models, this computes the element-wise mean for each
sub-kernel created by the iteration procedure.
}
