% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\value{
The value of the final expression in the pipe chain.
}
\description{
See vignette at \url{https://CRAN.R-project.org/package=magrittr}.
}
\keyword{internal}
