% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.matrix}
\alias{plot.matrix}
\alias{plot.simple_di_det_ipm}
\title{Plot a matrix or an *_ipm object}
\usage{
\method{plot}{matrix}(x = NULL, y = NULL, A, col = rainbow(100, start
  = 0.67, end = 0), bw = FALSE, do_contour = FALSE,
  do_legend = FALSE, ...)

\method{plot}{simple_di_det_ipm}(x = NULL, y = NULL, ipm,
  sub_kernels = FALSE, col = rainbow(100, start = 0.67, end = 0),
  bw = FALSE, do_contour = FALSE, do_legend = FALSE, ...)
}
\arguments{
\item{x, y}{Either the values of the meshpoints or \code{NULL}. If \code{NULL},
then a sequence is generated so that meshpoints are given sequential bin numbers.}

\item{A, ipm}{A matrix or a result from \code{make_ipm}}

\item{col}{A vector of colors to use for plotting}

\item{bw}{A logical indicating whether to use a greyscale palette for plotting}

\item{do_contour}{A logical indicating whether or not draw contour lines
on the plot}

\item{do_legend}{A logical indicating whether to draw a legend for the plot}

\item{...}{further arguments passed to legend}

\item{sub_kernels}{A logical - also plot the sub-kernels?}
}
\value{
\code{A} or \code{ipm} invisibly
}
\description{
Plot a matrix or an *_ipm object
}
