% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.ipmr_uncertainty}
\alias{plot.ipmr_uncertainty}
\title{Plot method for ipmr uncertainty objects}
\usage{
\method{plot}{ipmr_uncertainty}(x, type = c("all", "param", "vr", "stacked"), vr_colors = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{ipmr_uncertainty}}

\item{type}{Type of plot:
\itemize{
\item "param" - parameter-level 3-panel plot
\item "vr" - vital-rate aggregated contributions
\item "stacked" - stacked contributions vs lambda variance
\item "all" - returns all plots as a list
}}

\item{vr_colors}{A named character vector of colors used to represent
vital rate categories in plots. Names must match the values in the
\code{vital_rate} column of the uncertainty output. If \code{NULL},
a default colorblind-friendly palette is used.}

\item{...}{Additional arguments passed to plotting functions}
}
\value{
A \code{ggplot} object or a list of plots.
}
\description{
Plot method for ipmr uncertainty objects
}
