% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-export.R
\name{domains}
\alias{domains}
\alias{vital_rates}
\alias{kernel_formulae}
\title{Accessor functions for proto_ipm objects}
\usage{
domains(proto_ipm)

vital_rates(proto_ipm)

kernel_formulae(proto_ipm)
}
\arguments{
\item{proto_ipm}{A \code{proto_ipm} object.}
}
\value{
A named list of either numbers (for domains), or bare expressions (for
vital rates and kernel formulae).
}
\description{
Functions that access specific slots of a \code{proto_ipm} for
user examination.
}
