% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-generics.R
\name{print.proto_ipm}
\alias{print.proto_ipm}
\title{Generic methods for \code{proto_ipm}s}
\usage{
\method{print}{proto_ipm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{proto_ipm}}

\item{...}{Ignored}
}
\description{
Generic methods for \code{proto_ipm}s
}
