% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{qsd_converge}
\alias{qsd_converge}
\alias{qsd_converge.simple_di_stoch_kern_ipm}
\title{Compute the quasi-stable trait distribution for a stochastic model}
\usage{
qsd_converge(ipm, p_nms, tol, start_life, n_steps, ...)

\method{qsd_converge}{simple_di_stoch_kern_ipm}(
  ipm,
  p_nms,
  tol = 1e-07,
  start_life = rep(1, length(p_nms)),
  n_steps = 1000,
  ...
)
}
\arguments{
\item{ipm}{The output from \code{make_ipm}}

\item{p_nms}{The name of the survival/growth kernel(s)}

\item{tol}{The distance between each kernels true eigenvector and the
quasi-stable distribution that is acceptable. Computed as
\code{0.5 * sum(abs(pop_size - right_ev))}. Default is \code{1e-7}}

\item{start_life}{The index of the population vector to start on. Should
be a named list where names correspond to population states and entries
are length 1 integer vectors. Default is 1.}

\item{n_steps}{The number of times to iterate the model before checking
for convergence. Default is 1000.}

\item{...}{other arguments passed to methods.}
}
\description{
Compute the quasi-stable trait distribution for a stochastic model
}
