% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{iceplant_ex}
\alias{iceplant_ex}
\title{Raw demographic data to construct an example IPM}
\format{
288 observations of 10 variables
\describe{

  \item{id}{Individual identification number}
  \item{size}{Surface area in square meters of each individual at time \emph{t}.}
  \item{flower_n}{If the plant is reproductive, the number of flowers it made.}
  \item{log_size}{Log transformed \code{size}.}
  \item{repro}{Either 0 or 1 to indicate whether the plant is reproductive.}
  \item{size_next}{Surface area in square meters of each individual at time \emph{t + 1}.}
  \item{flower_n_next}{If the plant is reproductive at \emph{t + 1}, the number of
  flowers it made.}
  \item{survival}{Either 0 or 1 to indicate whether a plant at \emph{t} survives to \emph{t + 1}.}
  \item{log_size_next}{Log transformed \code{size_next}.}
  \item{repro_next}{Either 0 or 1 to indicate whether a plant is reproductive at \emph{t + 1}.}
}
}
\usage{
iceplant_ex
}
\description{
Raw demographic data to construct an example IPM
}
\keyword{datasets}
