% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-export.R
\name{right_mult}
\alias{right_mult}
\title{Right multiplication}
\usage{
right_mult(...)
}
\arguments{
\item{...}{Symbols representing functions - usually things that appear on the
left hand side of \code{formula} and/or \code{...} in \code{define_kernel} and
\code{define_k}.}
}
\description{
Performs right multiplication.
}
