% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{sim_di_det_ex}
\alias{sim_di_det_ex}
\title{Simple deterministic IPM example}
\format{
A simple deterministic IPM with the following slots:
\describe{

  \item{sub_kernels}{The computed sub-kernels, named \code{P} and \code{F}.}
  \item{env_list}{Empty}
  \item{env_seq}{Empty.}
  \item{pop_state}{Empty.}
  \item{proto_ipm}{The \code{proto_ipm} object used to implement the model.}

}
}
\usage{
sim_di_det_ex
}
\description{
Simple deterministic IPM example
}
\keyword{datasets}
