% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stoch.R
\name{mean_kernel}
\alias{mean_kernel}
\title{Helpers for stochastic models}
\usage{
mean_kernel(ipm)
}
\arguments{
\item{ipm}{A stochastic model created by \code{make_ipm()}.}
}
\value{
\code{mean_kernel} returns a list of mean sub-kernels
for the model.
}
\description{
Various functions to compute useful quantities for stochastic
models
}
