% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tidy-eval.R
\name{tidyeval}
\alias{tidyeval}
\alias{quo}
\alias{quos}
\alias{enquo}
\alias{enquos}
\alias{quo_name}
\alias{sym}
\alias{ensym}
\alias{syms}
\alias{ensyms}
\alias{expr}
\alias{exprs}
\alias{enexpr}
\alias{enexprs}
\alias{.data}
\title{Tidy eval helpers}
\description{
Re-exported tidy eval helpers
}
\details{
To learn more about tidy eval and how to use these tools, visit
\url{http://rlang.r-lib.org} and the \href{https://adv-r.hadley.nz/meta.html}{Metaprogramming section} of \href{https://adv-r.hadley.nz}{Advanced R}.
}
\keyword{internal}
