% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{update.proto_ipm}
\alias{update.proto_ipm}
\title{Update expressions in an IPM}
\usage{
\method{update}{proto_ipm}(object, kernels, formulae, dots, parameters, functions, ...)
}
\arguments{
\item{object}{A \code{proto_ipm} or output from \code{make_ipm}.}

\item{kernels}{The names of the kernels to modify.}

\item{formulae}{The new kernel formulae (if any). This should be a named expression
or list of expressions.}

\item{dots}{The new vital rate expressions (e.g. the \code{...} of \code{define_(k)ernel}).
This should be a named expression or list of expressions.}

\item{parameters}{A list of new parameters to include in the model. Names that
overlap with the current parameter list will overwrite the current values, while
names that are not in the current parameter list will be appended to it.}

\item{functions}{New user defined functions to include in the model.}

\item{...}{not used}
}
\description{
Provides methods to readily update vital rate functions and kernel
formulae, as well as pass new parameter values and/or functions to the model.
}
