% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-export.R
\name{vec}
\alias{vec}
\alias{inv_vec}
\title{Perform the \emph{vec} operation and it's inverse}
\usage{
vec(mat)

inv_vec(mat, square = TRUE, nrow = NULL, ncol = NULL)
}
\arguments{
\item{mat}{The matrix}

\item{square}{A logical indicating whether to return a square matrix. For IPMs,
this should usually be \code{TRUE}. If set to \code{FALSE}, \code{nrow} and \code{ncol}
must be specified}

\item{nrow}{The number of rows for the new matrix}

\item{ncol}{The number of columns for the new matrix}
}
\value{
\code{vec}: A matrix with dimensions nrow(mat) * ncol(mat) X 1

\code{inv_vec}: A matrix with dimensions \code{sqrt(length(mat))} X \code{sqrt(length(mat))} or
dimensions specified by \code{nrow} and \code{ncol}
}
\description{
Unstacks a matrix into a single column vector. Each successive
column is appended to the bottom of the vector
}
\examples{

mat <- matrix(1:20, nrow = 4, ncol = 5)

v <- vec(mat)

v

inv_vec(v)


}
