% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-export.R
\name{all_ages}
\alias{all_ages}
\title{Age X Size models}
\usage{
all_ages(expr, fun, ...)
}
\arguments{
\item{expr}{The expression to modify. See details.}

\item{fun}{The function to apply over \code{expr}.}

\item{...}{Used internally. Modifying this will likely cause models to break.}
}
\value{
A modified expression.
}
\description{
Helper functions for implementing age x size models
}
\details{
Currently only \code{all_ages} is exported. This is a helper
to indicate that all of the levels of age should be included in a given expression,
rather than split out into separate levels. The \code{fun} is used to combine
the results of each individual evaluation of \code{expr}.

The most common use case for \code{all_ages} will be when defining the size
distribution of new recruits in an age structured model, when many \code{F} kernels
are multiplied by their respective age X size distributions, and then summed.
Note that using \code{sum} is almost never correct, as this will always return
a single number whereas we typically want the trait distribution function
(i.e. a vector). It is usually best to use \code{+} instead.
}
\examples{

\dontrun{

define_k(
 proto_ipm          = some_age_size_proto_ipm,
 name               = "K",
 family             = "IPM",
 n_wt_0_t_1         = all_ages(expr = F_age \%*\% n_wt_age_t, fun = "+"),
 n_wt_age_t_1       = P_age_minus_1 \%*\% n_wt_age_minus_1_t,
 n_wt_max_age_t_1   = P_max_age \%*\% n_wt_max_age_t +
                      P_max_age_minus_1 \%*\% n_wt_max_age_minus_1_t,
 data_list          = param_list,
 states             = list (c("wt")),
 has_hier_effs      = FALSE,
 levels_ages        = list(age = c(0:20), max_age = 21),
 evict_cor          = FALSE
)


}

}
