% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-age_size.R
\name{sum.age_size_expr}
\alias{sum.age_size_expr}
\alias{prod.age_size_expr}
\title{Age X Size models}
\usage{
\method{sum}{age_size_expr}(expr, ..., na.rm = NA)

\method{prod}{age_size_expr}(expr, ..., na.rm = NA)
}
\arguments{
\item{expr}{An expression to compute a sum for.}

\item{...}{Used internally. Modifying this will likely cause models to break.}

\item{na.rm}{Not used and ignored.}
}
\value{
A modified expression.
}
\description{
Helper functions for implementing age x size models
}
\details{
Note that these functions do not compute a single number, but rather
expand an expression to compute the trait distribution. This differs from the
usual behavior of the numeric methods for \code{sum, prod} as those
return a single number or \code{NA}.
}
\examples{

\dontrun{

define_k(
 proto_ipm          = some_age_size_proto_ipm,
 name               = "K",
 family             = "IPM",
 n_wt_0_t_1         = sum(F_age \%*\% n_wt_age_t),
 n_wt_age_t_1       = P_age_minus_1 \%*\% n_wt_age_minus_1_t,
 n_wt_max_age_t_1   = P_max_age \%*\% n_wt_max_age_t +
                      P_max_age_minus_1 \%*\% n_wt_max_age_minus_1_t,
 data_list          = param_list,
 states             = list (c("wt")),
 has_hier_effs      = FALSE,
 levels_ages        = list(age = c(0:20), max_age = 21),
 evict_cor          = FALSE
)


}

}
