% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-export.R
\name{all_ages}
\alias{all_ages}
\title{Age X Size models}
\usage{
all_ages(expr, fun, ...)
}
\arguments{
\item{expr}{The expression to modify. See details.}

\item{fun}{The function to apply over \code{expr}.}

\item{...}{Used internally. Modifying this will likely cause models to break.}
}
\value{
A modified expression.
}
\description{
Helper functions for implementing age x size models
}
\details{
Currently only \code{all_ages} is exported. This is a helper
to indicate that all of the levels of age should be included in a given expression,
rather than split out into separate levels. The \code{fun} is used to combine
the results of each individual evaluation of \code{expr}.

The most common use case for \code{all_ages} will be when defining the size
distribution of new recruits in an age structured model, when many \code{F} kernels
are multiplied by their respective age X size distributions, and then summed.
Note that using \code{sum} is almost never correct, as this will always return
a single number whereas the size distribution should be a vector! Generally,
use \code{+} instead.
}
