% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-lin_alg.R
\name{is_conv_to_asymptotic}
\alias{is_conv_to_asymptotic}
\title{Check for model convergence to asymptotic dynamics}
\usage{
is_conv_to_asymptotic(ipm, tol = 1e-10)
}
\arguments{
\item{ipm}{An object returned by \code{make_ipm()}.}

\item{tol}{The tolerance for convergence. Convergence is evaluated by making an
element by element comparison  for the population state vectors at time \emph{t}
and \emph{t-1}.}
}
\value{
Either \code{TRUE} or \code{FALSE}.
}
\description{
Check for model convergence to asymptotic dynamics
}
