% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dom-states.R
\name{define_domains}
\alias{define_domains}
\title{Domains and State Variables}
\usage{
define_domains(proto_ipm, ...)
}
\arguments{
\item{proto_ipm}{The model to add the domain information to.}

\item{...}{Numeric vectors of length 3 where the name corresponds to the
state variable, the first entry is the lower bound of the domain, the second
is the upper bound of the domain, and the third entry is the number of
meshpoints.}
}
\value{
The \code{proto_ipm} with the domain information added.
}
\description{
Functions to assist with constructing and manipulating
the domains of state variables in \code{ipmr}.
}
