% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{right_ev}
\alias{right_ev}
\alias{right_ev.simple_di_det_ipm}
\alias{right_ev.general_di_det_ipm}
\alias{left_ev}
\title{Compute left and right eigenvectors via iteration}
\usage{
right_ev(ipm, ...)

\method{right_ev}{simple_di_det_ipm}(ipm, n_iterations = 100, ...)

\method{right_ev}{general_di_det_ipm}(ipm, n_iterations = 100, ...)

left_ev(ipm, ...)
}
\arguments{
\item{ipm}{Output from \code{make_ipm()}.}

\item{...}{other arguments passed to methods}

\item{n_iterations}{The number of times to iterate the model to reach
convergence. default is 100.}
}
\description{
Compute left and right eigenvectors via iteration
}
