% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{elasticity}
\alias{elasticity}
\alias{elasticity.simple_di_det_ipm}
\alias{elasticity.general_di_det_ipm}
\title{Compute elasticity}
\usage{
elasticity(
  ipm,
  what = c("r_0", "lambda", "gen_t"),
  level = c("kernel", "vital_rate", "parameter"),
  subset = NA_character_,
  ...
)

\method{elasticity}{simple_di_det_ipm}(
  ipm,
  what = "lambda",
  level = "kernel",
  subset = NA_character_,
  n_iterations = 100,
  ...
)

\method{elasticity}{general_di_det_ipm}(
  ipm,
  what = "lambda",
  level = "kernel",
  subset = NA_character_,
  n_iterations = 100,
  mega_mat,
  mega_vec,
  ...
)
}
\arguments{
\item{ipm}{Output from \code{make_ipm()}.}

\item{what}{The numerator of the partial derivative for sensitivity.
Possible options \code{"lambda"} (default), \code{"r_0"},
and \code{"gen_t"}. More will probably be added later.}

\item{level}{The level to compute elasticity at. \code{"kernel"} computes
the model wide elasticity surface and returns that. \code{"vital_rate"}
computes lambda's elasticity to specific vital rates, \code{"parameter"} lambda's
elasticity to each parameter. Use \code{subset} to specify a subset of
vital rates or parameters to compute values for.}

\item{subset}{A character vector corresponding to the \code{"vital_rates"} or
\code{"parameters"} you want to compute  sensitivities for. Can save time if
only a few parameters or vital rates are of interest.}

\item{...}{other arguments passed to methods}

\item{n_iterations}{The number of times to iterate the model when computing
the left and right eigenvectors. Default is 100 for deterministic IPMs, and varies
depending on the stochastic IPM.}

\item{mega_mat}{A vector of names and/or 0s specifying the relationship
between the kernels in the model. The names should correspond to kernel
names, with 0s corresponding to sparse areas of the mega-matrix. The names can
be either symbols or characters. These functions support suffix expansion as in
\code{define_k(ernel)}, so expressions don't need to be re-written for every
combination hierarchical effects.}

\item{mega_vec}{A vector of names specifying the format of the population
vector. The names can be either symbols or characters.}
}
\value{
A list. Contains either kernel response surfaces, vital rate level
elasticity values, or parameter level elasticity values. The class is always
\code{c("ipmr_elasticity", "list")} (for internal usage and plot methods).
}
\description{
Compute elasticity
}
