% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eviction.R
\name{truncated_distributions}
\alias{truncated_distributions}
\title{Eviction correction}
\usage{
truncated_distributions(fun, param, L = NA, U = NA, ...)
}
\arguments{
\item{fun}{The cumulative density function to use. For example, could be
\code{"norm"} to correct a Gaussian density function, or \code{"lnorm"} to
correct a log-normal density function.}

\item{param}{The parameter/vital rate being modified. If this is a vector, the
distribution specified in \code{fun} will be recycled.}

\item{L}{Optionally, the name of the lower bound of the domain. Otherwise,
the function will try to find the \code{param} inside of the \code{proto_ipm}
object under construction and use that to infer the domain of the function.}

\item{U}{Optionally, the name of the upper bound of the domain. Otherwise,
the function will try to find the \code{param} inside of the \code{proto_ipm}
object under construction and use that to infer the domain of the function.}

\item{...}{Only used for internal modification - do not use!}
}
\value{
A modified function call with that re-scales the probability density
function based on the cumulative density function.
}
\description{
Various helpers to correct for unintentional eviction (Williams
et al. 2012). \code{rescale_kernel} is an alias for
\code{truncated_distributions}.
}
\references{
Williams JL, Miller TEX & Ellner SP, (2012). Avoiding unintentional eviction
from integral projection models.Ecology 93(9): 2008-2014.
}
