% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eviction.R
\name{truncated_distributions}
\alias{truncated_distributions}
\title{Eviction correction}
\usage{
truncated_distributions(discretized_kernel)
}
\arguments{
\item{discretized_mat}{The kernel or function that needs correcting.}
}
\value{
A matrix of the same dimension as the input.
}
\description{
Various helpers to correct for unintentional eviction (Williams
et al. 2012).
}
\references{
Williams JL, Miller TEX & Ellner SP, (2012). Avoiding unintentional eviction
from integral projection models.Ecology 93(9): 2008-2014.
}
