% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-export.R
\name{format_mega_matrix}
\alias{format_mega_matrix}
\title{Format a mega-matrix}
\usage{
format_mega_matrix(ipm, mega_mat, presets = NULL)
}
\arguments{
\item{ipm}{Output from \code{make_ipm}.}

\item{mega_mat}{A vector with symbols and/or 0s representing the matrix blocks.
They should be specified in ROW MAJOR order! See examples.}

\item{presets}{Either empty or one of 'age-size' or 'hier_effs'. Currently
not implemented}
}
\value{
A large matrix
}
\description{
Format a mega-matrix
}
\examples{
data(gen_di_det_ex)

big_k <- format_mega_matrix(gen_di_det_ex,
                            mega_mat = c(0, go_discrete,
                                         leave_discrete, P))



}
