% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ipm_report.R
\name{make_ipm_report}
\alias{make_ipm_report}
\alias{make_ipm_report_body}
\title{Generate an RMarkdown file with IPM metadata}
\usage{
make_ipm_report(
  proto_ipm,
  rmd_dest = getwd(),
  title = "",
  output_format = "html",
  render_output = FALSE,
  translate_greek = TRUE
)

make_ipm_report_body(proto_ipm, translate_greek)
}
\arguments{
\item{proto_ipm}{A proto_ipm object}

\item{rmd_dest}{The folder to save the Rmd file at. The default is
\code{getwd()}. Alternatively, can be a complete file path that specifies
the location and title of the document with the extension \code{".rmd"}. in
this case, the current date will be appended to the title.}

\item{title}{The title to include in the document. This is not necessarily
the same as \code{rmd_dest}, as this appears at the top of the generated
report, and is not included in the file path!}

\item{output_format}{The format to include in the YAML header for the created
\code{.rmd} document.}

\item{render_output}{A logical indicating whether to call
\code{rmarkdown::render} on the generated \code{.rmd} file. Often times, the
\code{.rmd} file will need further editing before it's useful, so the default
is \code{FALSE}.}

\item{translate_greek}{A logical. \code{TRUE} means the function will attempt
to translate symbols in vital rate expressions to Greek equivalents, which
can make reports prettier. \code{FALSE} will leave the expressions largely
as is before translating them to LaTeX equivalents.}
}
\value{
For \code{make_ipm_report}, the filepath to the \code{.rmd} file. The
default name is \code{ "ipmr_report_<current_date>.rmd"}. For
\code{make_ipm_report_body}, a character vector with Markdown and LaTeX
suitable for rendering, but without a header.
}
\description{
Generates a \code{.rmd} file containing a mathematical description
of the \code{proto_ipm} object.
}
\details{
\code{make_ipm_report_body} only translates the iteration
  expressions and vital rate expressions into Markdown with LaTeX, and does
  not produce any headers needed to knit the file. This function is exported
  mostly for re-usage in \code{\link[Rpadrino]{pdb_report}}, and isn't really
  intended for use by \code{ipmr} users.

  @section \strong{Translations}

  For iteration expressions, vital rate expressions, and parameter names,
  \code{make_ipm_report} first attempts to translate functional forms into a
  common set of names. For example, \code{s = surv_int + surv_slope * z_1} is
  translated into \code{beta_0_s + beta_1_s * z_1}, and then is
  translated into LaTeX equations. This helps create useful Greek symbols
  more consistently. This may cause strange/incorrect results when parsing
  non-linear models.
}
