% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_kernel.R
\name{define_kernel}
\alias{define_kernel}
\alias{define_k}
\title{Functions to initialize and define IPM kernels}
\usage{
define_kernel(proto_ipm, name, formula, family, ..., data_list = list(),
  state_list, has_hier_effs = FALSE, levels_hier_effs = list(),
  evict = TRUE, evict_fun = NULL)

define_k(proto_ipm, name, formula, family, ..., data_list = list(),
  state_list, has_hier_effs = FALSE, levels_hier_effs = list(),
  evict = TRUE, evict_fun = NULL)
}
\arguments{
\item{proto_ipm}{The name of object you wish to append the new kernel to.}

\item{name}{The name of the new kernel. Survival and growth kernels should be
named \code{"P"} for now. Fecundity should be named \code{"F"}, and clonal
kernels should be named \code{"C"} (again, just for now. \strong{ need to work on
generalizing that).}}

\item{formula}{An bare expression specifying the form of the kernel. See below
for examples}

\item{family}{The type of kernel. Options are \code{"CC"} for continuous to continuous
transitions, \code{"DC"} for discrete to continuous (e.g. emergence from a seedbank),
\code{"CD"} for continuous to discrete (e.g. entering a seedbank), and \code{"DD"} for
discrete to discrete (e.g. stasis in a seedbank).}

\item{...}{A set of named expressions that correspond to vital rates in \code{formula}.}

\item{data_list}{A list of named values that correspond to constants in the formula.
You do not need to specify vectors corresponding to the domains here.}

\item{state_list}{A list containing the names of each state variable used in
the kernel.}

\item{has_hier_effs}{A logical indicating whether or not the kernel and/or its
underlying vital rates are structured with hierarchical effects. If so and you
specify either the functional forms or the exact parameter values you want to use,
the kernels will be automatically split according to the notation in the vital
rates and kernel formulae and multiple kernels will be built. See the vignette
on the syntax for this feature for more details (\code{vignettes(
'hierarchical-notation', package = 'ipmr')}).}

\item{levels_hier_effs}{A named list with vectors corresponding the various levels
the hierarchical variable can take. This should be a single vector, usually
character or integer.}

\item{evict}{A logical indicating whether an eviction correction should be applied
to the kernel. Default is \code{TRUE}.}

\item{evict_fun}{If \code{evict == TRUE}, then a function that corrects for it.
Currently, \code{truncated_distributions} and \code{rescale_kernel} are the
only built-in functions, but user specified ones should work as well.}
}
\value{
All functions described here return a \code{proto_ipm}.
}
\description{
Adds a new kernel to the \code{proto_ipm} structure. Different
classes of IPMs may have many or only a few kernels. Each one requires its
own call to this function, though there are some exceptions, namely for kernels
derived from hierarchical models (e.g. vital rate models fit across plots and years).
See Details for more info
}
\details{
\strong{BLAH BLAH BLAH}
}
