% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_kernel.R
\name{define_kernel}
\alias{define_kernel}
\alias{define_k}
\alias{define_k.default}
\alias{define_k.age_x_size}
\alias{remove_k}
\title{Functions to initialize and define IPM kernels}
\usage{
define_kernel(
  proto_ipm,
  name,
  formula,
  family,
  ...,
  data_list = list(),
  states,
  has_hier_effs = FALSE,
  levels_hier_effs = list(),
  levels_ages = list(),
  evict_cor = FALSE,
  evict_fun = NULL
)

define_k(
  proto_ipm,
  name,
  family,
  ...,
  data_list = list(),
  states,
  has_hier_effs = FALSE,
  levels_hier_effs = list(),
  levels_ages = list(),
  evict_cor = FALSE,
  evict_fun = NULL
)

\method{define_k}{default}(
  proto_ipm,
  name,
  family,
  ...,
  data_list = list(),
  states,
  has_hier_effs = FALSE,
  levels_hier_effs = list(),
  evict_cor = FALSE,
  evict_fun = NULL
)

\method{define_k}{age_x_size}(
  proto_ipm,
  name,
  family,
  ...,
  data_list = list(),
  states,
  has_hier_effs = FALSE,
  levels_hier_effs = list(),
  levels_ages = list(),
  evict_cor = FALSE,
  evict_fun = NULL
)

remove_k(proto_ipm)
}
\arguments{
\item{proto_ipm}{The name of the model.}

\item{name}{The name of the new kernel. There are only two rules: 1. For
\code{define_k}, the name must start with \code{"K"} or \code{"k"}.
2. For \code{define_kernel}, the name cannot start with \code{"K"} or
\code{"k"}.}

\item{formula}{A bare expression specifying the form of the kernel. See Details.}

\item{family}{The type of kernel. Options are \code{"CC"} for continuous to continuous
transitions, \code{"DC"} for discrete to continuous (e.g. emergence from a seedbank),
\code{"CD"} for continuous to discrete (e.g. entering a seedbank), and \code{"DD"} for
discrete to discrete (e.g. stasis in a seedbank).}

\item{...}{For \code{define_kernel}, a set of named expressions that correspond
to vital rates in \code{formula}. For \code{define_k}, a set of named expressions
that relate the population state at T + 1 to the population state at T. Alternatively,
can be an expression where the left hand side is the name of the kernel specified
in \code{name} and the right hand side only describes the structure of the
iteration kernel. In all cases, suffix expansion of hierarchical models is supported.
Specific details for each case are in their respecitive function's sections
below.}

\item{data_list}{A list of named values that correspond to constants in the formula.
You do not need to specify vectors corresponding to the domains here.}

\item{states}{A list with character vector containing the names of each state
variable used in the kernel.}

\item{has_hier_effs}{A logical indicating whether or not the kernel and/or its
underlying vital rates are structured with hierarchical effects. If so and you
specify either the functional forms or the exact parameter values you want to use,
the kernels will be automatically split according to the notation in the vital
rates and kernel formulae and multiple kernels will be built. See the vignette
on the syntax for this feature for more details (\code{vignettes(
'hierarchical-notation', package = 'ipmr')}).}

\item{levels_hier_effs}{A named list with vectors corresponding to the various levels
the hierarchical variable can take. The names should match the suffixes used
in the vital rate expressions.}

\item{levels_ages}{A list with possible two entries: 1. \code{"age"}: the range
of ages, and, optionally, 2. \code{"max_age"}: The maximum age for the model.}

\item{evict_cor}{A logical indicating whether an eviction correction should be applied
to the kernel.}

\item{evict_fun}{If \code{evict_cor== TRUE}, then a function that corrects for it.
Currently, the only implemented function is \code{truncated_distributions}.
This works by modifying the functional form of the vital rate expressions
\code{...}, and so supplying your own function in this slot will be difficult.
You can also either specify a \code{usr_funs} function that performs the correction
during the implementation of the model itself, or incorporate the correction
into individual vital rate expressions. In either of those cases,
set \code{evict_cor} to \code{FALSE}.}
}
\value{
All functions described here return a  \code{proto_ipm}.
}
\description{
Adds a new kernel to the \code{proto_ipm} structure.
}
\details{
Different classes of IPMs may have many or only a few kernels. Each
one requires its own call to this function, though there are some exceptions,
namely for kernels derived from hierarchical models (e.g. vital rate models
fit across plots and years). \code{remove_k} is a helper for re-building models
but with a different \code{K} kernel.

Internally, \code{ipmr} generates the \code{levels_hier_effs} using \code{expand.grid}.
This  means that every combination of the values in \code{levels_hier_effs}
must exist in the \code{data_list}, or the model will fail with an error along
the lines of \code{Error in eval_tidy: object 'x_yz' not found}. In order to
exclude levels that do not exist in your data, you can add a vector to the list
in \code{levels_hier_effs} called \code{drop_levels}. This should contain the
values you wish to exclude as a character vector. For example, if data are
collected for \code{sites = c("a", "b", "c")}, and \code{years = c(2005:2008)},
but there is no data from site \code{"a"} in \code{2007}, we can use
\code{levels_hier_effs = list(site = c("a", "b", "c"), year = c(2005:2008),
drop_levels = c("a_2007"))}.

\strong{\code{define_k}}

The preferred method of defining a \code{K} kernel is to use the left
hand side of the \code{...} to reference the population vectors that the right
hand side creates (e.g. \code{n_T_1 = (P + F) \%*\% n_T}). This enables
iteration-based methods to work properly. On the other hand, these iterations
can be quite time consuming, and some applications only require
an iteration matrix while not necessarily requiring the population vectors (e.g.
calculations of deterministic population growth rate). In those cases, the
\code{...} can contain something like \code{K = P + F}. In this case, the left
hand side of the expression should match the \code{name} argument to \code{define_k}.

\strong{\code{define_kernel}}

\code{define_kernel} generates most of the information needed to create an IPM
kernel. There are a few requirements - \code{name}, \code{family},
\code{formula}, and \code{states} must not be empty. The \code{formula} should be an expression
for how vital rates produce a kernel (e.g. \code{formula = S * G}). The
\code{...} should be a set of named expressions that correspond to vital rate
expressions (e.g. \code{G = dnorm(size_2, mean_size, sd_size)}). See the vignettes
for more information on how to get started.
}
