% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{lambda}
\alias{lambda}
\alias{lambda.simple_di_det_ipm}
\alias{lambda.simple_di_stoch_kern_ipm}
\alias{lambda.simple_di_stoch_param_ipm}
\alias{lambda.general_di_det_ipm}
\alias{lambda.general_di_stoch_kern_ipm}
\alias{lambda.general_di_stoch_param_ipm}
\title{Compute the per-capita growth rate for an IPM object}
\usage{
lambda(ipm, ...)

\method{lambda}{simple_di_det_ipm}(ipm, comp_method = "pop_size", type_lambda = "last", ...)

\method{lambda}{simple_di_stoch_kern_ipm}(ipm, comp_method = "pop_size", type_lambda = "stochastic", ...)

\method{lambda}{simple_di_stoch_param_ipm}(ipm, comp_method = "pop_size", type_lambda = "stochastic", ...)

\method{lambda}{general_di_det_ipm}(ipm, comp_method = "pop_size", type_lambda = "last", ...)

\method{lambda}{general_di_stoch_kern_ipm}(ipm, ..., comp_method = "pop_size", type_lambda = "stochastic")

\method{lambda}{general_di_stoch_param_ipm}(ipm, ..., comp_method = "pop_size", type_lambda = "stochastic")
}
\arguments{
\item{ipm}{An object returned by \code{make_ipm()}.}

\item{...}{other arguments passed to methods.}

\item{comp_method}{Either \code{"eigen"} or \code{"pop_size"}. \code{"eigen"}
is only possible for \code{"simple_*"} methods.}

\item{type_lambda}{Either \code{'all'}, \code{'last'},
 or \code{'stochastic'}. \code{'all'}
returns a vector of lambda values for each time step of the simulation (equal
in length to the \code{iterations} argument of \code{make_ipm()}).
\code{'last'} returns the lambda value for the final timestep.
\code{'stochastic'} returns a single value which is the geometric mean
of log-per-capita growth rate from each time step.}
}
\value{
An array. Rows correspond to time steps, and columns correspond
to hierarchical effects (if any).
}
\description{
Compute the per-capita growth rate for an IPM object
}
\details{
There are two possible methods for computing \code{lambda} and these
are controlled by the \code{comp_method} argument. Possible values are
\code{"eigen"} and \code{"pop_size"}. The first computes the dominant
eigenvalue of all entries in \code{ipm$iterators} slot of the \code{*_ipm}
object. Since iteration kernels aren't generated for \code{general_*} methods,
\code{"pop_size"} is the only possible option for those objects.
}
