% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ipm.R
\name{make_ipm}
\alias{make_ipm}
\title{Methods to implement an IPM}
\usage{
make_ipm(proto_ipm, ...)
}
\arguments{
\item{proto_ipm}{The proto_ipm object you wish to implement. This should be the
output of \code{add_kernel}, \code{add_K}, or the \code{define_*} functions.}

\item{...}{Other arguments passed to methods}
}
\value{
The \code{make_ipm.*det} methods will always return a list of length 4
containing the following components:

\itemize{
  \item{\strong{iterators}}{: iteration kernel(s) (if specified by \code{add_K}),
                            otherwise contains \code{NA}.}
  \item{\strong{sub_kernels}}{: the sub_kernels specified in \code{add_kernel}.}
  \item{\strong{pop_state}}{: population vectors stored as an instance of the
                             \code{pop_state} class.}
  \item{\strong{proto_ipm}}{: the \code{proto_ipm} object used to implement
                             the model.}
}
 The \code{make_ipm.*stoch} methods will always return a list of length 5
containing the following components:

\itemize{
  \item{\strong{iterators}}{: iteration kernel(s) (if specified by \code{add_K}),
                            otherwise contains \code{NA}.}
  \item{\strong{sub_kernels}}{: the sub_kernels specified in \code{add_kernel}.}
  \item{\strong{pop_state}}{: population vectors stored as an instance of the
                             \code{pop_state} class.}
  \item{\strong{env_state}}{: a matrix with dimension \code{n_iterations} X 1 of
                             kernel indices indicating the order
                             in which kernels are to be/were resampled OR
                             a matrix with as many columns as stochastic parameters
                             \code{n_iterations} rows.}
  \item{\strong{proto_ipm}}{: the \code{proto_ipm} object used to implement
                             the model.}
}
}
\description{
The \code{make_ipm.*} methods convert a \code{proto_ipm} into a
set of discretized kernels and population vectors. Methods have different
requirements, so carefully read the parameter documentation and the
\code{vignette("implementation", package = "ipmr")}.
}
\author{
Sam Levin
}
