% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.ipmr_matrix}
\alias{plot.ipmr_matrix}
\alias{plot.simple_di_det_ipm}
\alias{plot.simple_di_stoch_param_ipm}
\alias{plot.simple_di_stoch_kern_ipm}
\alias{plot.general_di_det_ipm}
\title{Plot a matrix or an *_ipm object}
\usage{
\method{plot}{ipmr_matrix}(
  x = NULL,
  y = NULL,
  A,
  col = grDevices::rainbow(100, start = 0.67, end = 0),
  bw = FALSE,
  do_contour = FALSE,
  do_legend = FALSE,
  ...
)

\method{plot}{simple_di_det_ipm}(
  x = NULL,
  y = NULL,
  ipm = NULL,
  col = rainbow(100, start = 0.67, end = 0),
  bw = FALSE,
  do_contour = FALSE,
  do_legend = FALSE,
  exponent = 1,
  ...
)

\method{plot}{simple_di_stoch_param_ipm}(
  x = NULL,
  y = NULL,
  ipm = NULL,
  col = rainbow(100, start = 0.67, end = 0),
  bw = FALSE,
  do_contour = FALSE,
  do_legend = FALSE,
  exponent = 1,
  ...
)

\method{plot}{simple_di_stoch_kern_ipm}(
  x = NULL,
  y = NULL,
  ipm = NULL,
  col = rainbow(100, start = 0.67, end = 0),
  bw = FALSE,
  do_contour = FALSE,
  do_legend = FALSE,
  exponent = 1,
  ...
)

\method{plot}{general_di_det_ipm}(
  x = NULL,
  y = NULL,
  ipm = NULL,
  mega_mat = NA_character_,
  col = NA_character_,
  bw = FALSE,
  do_contour = FALSE,
  do_legend = FALSE,
  exponent = 1,
  ...
)
}
\arguments{
\item{x, y}{Either the values of the meshpoints or \code{NULL}. If \code{NULL},
then a sequence is generated so that meshpoints are given sequential bin numbers.}

\item{A, ipm}{A matrix or a result from \code{make_ipm}, or \code{NULL} if \code{x}
is specified as the matrix or IPM object.}

\item{col}{A vector of colors to use for plotting}

\item{bw}{A logical indicating whether to use a greyscale palette for plotting}

\item{do_contour}{A logical indicating whether or not draw contour lines
on the plot}

\item{do_legend}{A logical indicating whether to draw a legend for the plot}

\item{...}{further arguments passed to legend}

\item{exponent}{The exponent to raise each kernel to. Setting this to a low
number can help visualize kernels that are overwhelmed by a few very large numbers.}

\item{mega_mat}{A vector with symbols, I's, and/or 0s representing the matrix blocks.
They should be specified in ROW MAJOR order! Can also be a character
string specifying the call. Hierarchical syntax is supported. When used,
\code{format_mega_matrix} will produce as many mega-matrices as there are
combinations of \code{levels_hier_effs} in the \code{proto_ipm}.}
}
\value{
\code{A} or \code{ipm} invisibly
}
\description{
Plot a matrix or an *_ipm object
}
\details{
\code{plot.ipmr_matrix} is intended for internal use only, and it
is usually safer to use \code{plot.*_ipm} methods for visualizing kernels.
If a K kernel is overwhelmed by information in say, a fecundity sub-kernel,
use the \code{exponent} argument in \code{plot.*_ipm} to make it more visually
appealing.
}
