% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.ipmr_uncertainty}
\alias{print.ipmr_uncertainty}
\title{Print method for ipmr uncertainty objects}
\usage{
\method{print}{ipmr_uncertainty}(x, ...)
}
\arguments{
\item{x}{An object of class `ipmr_uncertainty`}

\item{...}{Unused}
}
\value{
\code{x} invisibly.
}
\description{
Print method for ipmr uncertainty objects
}
