% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.proto_ipm}
\alias{print.proto_ipm}
\title{Methods for proto_ipms}
\usage{
\method{print}{proto_ipm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{proto_ipm}}

\item{...}{Ignored}
}
\description{
Methods for proto_ipms
}
