% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-export.R
\name{right_mult}
\alias{right_mult}
\alias{left_mult}
\title{Right/left multiplication}
\usage{
right_mult(kernel, vectr)

left_mult(kernel, vectr)
}
\arguments{
\item{kernel, vectr}{\code{kernel} should be a bivariate kernel, \code{vectr}
should be a univariate trait distribution.}
}
\value{
\code{left_mult} returns \code{t(kernel) \%*\% vectr}. \code{right_mult}
returns \code{kernel \%*\% vectr}.
}
\description{
Performs right and left multiplication.
}
