% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{sensitivity}
\alias{sensitivity}
\title{Compute sensitivity}
\usage{
sensitivity(
  ipm,
  level = c("kernel", "vital_rate", "parameter"),
  subset = NA_character_,
  ...
)
}
\arguments{
\item{ipm}{Output from \code{make_ipm()}.}

\item{level}{The level to compute sensitivity at. \code{"kernel"} computes
the model wide sensitivity surface and returns that. \code{"vital_rate"}
computes lambda's sensitivity to specific vital rates, \code{"parameter"} lambda's
sensitivity to each parameter. Use \code{subset} to specify a subset of
vital rates or parameters to compute values for.}

\item{subset}{A character vector corresponding to the \code{"vital_rates"} or
\code{"parameters"} you want to compute  sensitivities for. Can save time if
only a few parameters or vital rates are of interest.}

\item{...}{other arguments passed to methods}
}
\value{
A list. Contains either kernel response surfaces, vital rate level
sensitivity values, or parameter level sensitivity values. The class is always
\code{c("ipmr_sensitivity", "list")} (for internal usage and plot methods).
}
\description{
Compute sensitivity
}
