% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{sensitivity}
\alias{sensitivity}
\alias{sensitivity.simple_di_det_ipm}
\alias{sensitivity.general_di_det_ipm}
\title{Compute sensitivity}
\usage{
sensitivity(
  ipm,
  what = c("r_0", "lambda", "gen_t"),
  level = c("kernel", "vital_rate", "parameter"),
  subset = NA_character_,
  ...
)

\method{sensitivity}{simple_di_det_ipm}(
  ipm,
  what = "lambda",
  level = "kernel",
  subset = NA_character_,
  n_iterations = 100,
  ...
)

\method{sensitivity}{general_di_det_ipm}(
  ipm,
  what = "lambda",
  level = "kernel",
  subset = NA_character_,
  n_iterations = 100L,
  mega_mat,
  mega_vec,
  ...
)
}
\arguments{
\item{ipm}{Output from \code{make_ipm()}.}

\item{what}{The numerator of the partial derivative for sensitivity.
Possible options \code{"lambda"} (default), \code{"r_0"},
and \code{"gen_t"}. More will probably be added later.}

\item{level}{The denominator of the partial derivative for sensitivity.
\code{"kernel"} computes the model wide sensitivity surface and returns that.
\code{"vital_rate"} computes lambda's sensitivity to specific vital rates,
\code{"parameter"} lambda's sensitivity to each parameter. Use \code{subset}
to specify a subset of vital rates or parameters to compute values for.}

\item{subset}{A character vector corresponding to the \code{"vital_rates"} or
\code{"parameters"} you want to compute  sensitivities for. Can save time if
only a few parameters or vital rates are of interest.}

\item{...}{other arguments passed to methods}

\item{n_iterations}{The number of times to iterate the model when computing
the left and right eigenvectors. Default is 100 for deterministic IPMs, and varies
depending on the stochastic IPM.}

\item{mega_mat}{A vector of names and/or 0s specifying the relationship
between the kernels in the model. The names should correspond to kernel
names, with 0s corresponding to sparse areas of the mega-matrix. The names can
be either symbols or characters. These functions support suffix expansion as in
\code{define_k(ernel)}, so expressions don't need to be re-written for every
combination hierarchical effects.}

\item{mega_vec}{A vector of names specifying the format of the population
vector. The names can be either symbols or characters.}
}
\value{
A list. Contains either kernel response surfaces, vital rate level
sensitivity values, or parameter level sensitivity values. The class is always
\code{c("ipmr_sensitivity", "list")} (for internal usage and plot methods).
}
\description{
Compute sensitivity
}
